# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "cristinaperezboix.alu@iespacomolla.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/profile/edit": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Muestra el formulario de edición del perfil de usuario",
                "description": "Muestra el formulario de edición del perfil de usuario.",
                "operationId": "edit",
                "responses": {
                    "200": {
                        "description": "Devuelve la vista con el formulario de edición del perfil de usuario"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/update": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Actualiza el perfil de usuario",
                "description": "Actualiza el perfil de usuario con los datos proporcionados en la solicitud.",
                "operationId": "update",
                "requestBody": {},
                "responses": {
                    "302": {
                        "description": "Redirecciona al formulario de edición del perfil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/delete": {
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Elimina la cuenta de usuario",
                "description": "Elimina permanentemente la cuenta de usuario actual.",
                "operationId": "destroyProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "Contraseña actual del usuario",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirecciona a la página de inicio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Contraseña incorrecta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The password is incorrect."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}